Fixed a few compiler warnings.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/asyn/current@13691 ececa60f-9f63-0410-97c7-ea05888eb260

+++ asyn/vxi11/drvVxi11.c
@@ -1140,9 +1140,10 @@ static asynStatus vxiRead(void *drvPvt,asynUser *pasynUser,
             clntStat = clientIoCall(pvxiPort, pasynUser, device_read,
                 (const xdrproc_t) xdr_Device_ReadParms,(void *) &devReadP,
                 (const xdrproc_t) xdr_Device_ReadResp,(void *) &devReadR);
-            if(devReadP.io_timeout!=ULONG_MAX
-            || devReadR.error!=VXI_IOTIMEOUT
-            || devReadR.data.data_len>0) break;
+            if(	devReadP.io_timeout		!= UINT_MAX
+            ||	devReadR.error			!= VXI_IOTIMEOUT
+            ||	devReadR.data.data_len	> 0 )
+				break;
         }
         if(clntStat != RPC_SUCCESS) {
             epicsSnprintf(pasynUser->errorMessage,pasynUser->errorMessageSize,
